ref: refactors config
