+++ doc/modules/colorizer.html
@@ -253,7 +253,7 @@ Detach
     <h3>Parameters:</h3>
     <ul>
         <li><span class="parameter">bufnr</span>
-         number|nil: A value of 0 implies the current buffer.
+         number|nil: buffer number (0 for current)
         </li>
     </ul>
